requirements.txt
streamlit==1.39.0
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.0
plotly==5.24.1
google-generativeai==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1